+++ orfwx/R/getupdatedcsv.R
@@ -40,6 +40,16 @@ getUpdatedCsv <- function(){
       stop("Directory does not exist and apprently could not be created.")
     }
     utils::download.file(gucURL, fLocation)
+    # Notify the user whether or not the update is up-to-date or not
+    lastUpdated <- orfwx::findMostRecentDate()
+    if(lastUpdated == orfwx::yesterdate()) {
+      message(paste0("UP TO DATE! Updated through yesterday (", 
+                     lastUpdated, 
+                     ")"))
+    } else {
+      message(paste("OUT OF DATE\n-----------\nMost recent date included:",
+                    lastUpdated))
+    }
   } else {
     stop(paste("Update not written to file: A file at",
                fLocation,